+++ src/tests/parse/test_vla_analyzer.h
@@ -46,7 +46,7 @@ TEST test_scan_for_vlas_basic(void) {
                     "  int const c_vla[n];\n"
                     "}\n";
 
-  ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src), &tokens));
+  ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)(size_t)src), &tokens));
   (void)vla_site_list_init(&list);
 
   ASSERT_EQ(0, scan_for_vlas(tokens, &list));
@@ -125,7 +125,7 @@ TEST test_scan_for_vlas_oom(void) {
   int i;
   int res;
 
-  ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src), &tokens));
+  ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)(size_t)src), &tokens));
 
   for (i = 0; i < 25; ++i) {
     (void)vla_site_list_init(&list);
@@ -149,7 +149,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
 
   /* Empty array size */
   const char *src1 = "int arr[]; int arr2[ ];";
-  ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src1), &tokens));
+  ASSERT_EQ(0,
+            tokenize(az_span_create_from_str((char *)(size_t)src1), &tokens));
   (void)vla_site_list_init(&list);
   ASSERT_EQ(0, scan_for_vlas(tokens, &list));
   ASSERT_EQ(0, list.count); /* Not a VLA */
@@ -161,7 +162,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
    */
   {
     const char *src2 = "MyType arr[n]; MyType *arr2[n]; MyType   arr3[n];";
-    ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src2), &tokens));
+    ASSERT_EQ(0,
+              tokenize(az_span_create_from_str((char *)(size_t)src2), &tokens));
     (void)vla_site_list_init(&list);
     ASSERT_EQ(0, scan_for_vlas(tokens, &list));
     ASSERT_EQ(3, list.count);
@@ -172,7 +174,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
     /* Missing variable name (syntax error but we should not crash) */
     {
       const char *src3 = "int [n];";
-      ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src3), &tokens));
+      ASSERT_EQ(
+          0, tokenize(az_span_create_from_str((char *)(size_t)src3), &tokens));
       (void)vla_site_list_init(&list);
       ASSERT_EQ(0, scan_for_vlas(tokens, &list));
       vla_site_list_free(&list);
@@ -182,7 +185,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
       /* Missing semicolon */
       {
         const char *src4 = "int arr[n]";
-        ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)src4), &tokens));
+        ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)(size_t)src4),
+                              &tokens));
         (void)vla_site_list_init(&list);
         ASSERT_EQ(0, scan_for_vlas(tokens, &list));
         vla_site_list_free(&list);
@@ -192,8 +196,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
         /* Missing array bracket closing */
         {
           const char *src5 = "int arr[n";
-          ASSERT_EQ(0,
-                    tokenize(az_span_create_from_str((char *)src5), &tokens));
+          ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)(size_t)src5),
+                                &tokens));
           (void)vla_site_list_init(&list);
           ASSERT_EQ(0, scan_for_vlas(tokens, &list));
           vla_site_list_free(&list);
@@ -203,8 +207,8 @@ TEST test_scan_for_vlas_edge_cases(void) {
           /* Type modifiers with whitespace */
           {
             const char *src6 = "const  int  arr [ n ] ;";
-            ASSERT_EQ(0,
-                      tokenize(az_span_create_from_str((char *)src6), &tokens));
+            ASSERT_EQ(0, tokenize(az_span_create_from_str((char *)(size_t)src6),
+                                  &tokens));
             (void)vla_site_list_init(&list);
             ASSERT_EQ(0, scan_for_vlas(tokens, &list));
             ASSERT_EQ(1, list.count);
@@ -217,8 +221,9 @@ TEST test_scan_for_vlas_edge_cases(void) {
               const char *src7 =
                   "void f(int n) { int a[n]; int b[n]; int c[n]; int d[n]; "
                   "int e[n]; int f[n]; }";
-              ASSERT_EQ(
-                  0, tokenize(az_span_create_from_str((char *)src7), &tokens));
+              ASSERT_EQ(0,
+                        tokenize(az_span_create_from_str((char *)(size_t)src7),
+                                 &tokens));
               (void)vla_site_list_init(&list);
               ASSERT_EQ(0, scan_for_vlas(tokens, &list));
               ASSERT_EQ(6, list.count);
